requirements.txt
opencv-python
Pillow
kraken
pytesseract
